+++ src/lsapy/__init__.py
@@ -2,7 +2,7 @@
 
 from lsapy.criteria import SuitabilityCriteria
 from lsapy.functions import *
-from lsapy.lsa import LandSuitability
+from lsapy.lsa import LandSuitabilityAnalysis
 
 __author__ = "Baptiste Hamon"
 __email__ = "baptiste.hamon@pg.canterbury.ac.nz"